requirements.txt
# Django core
Django>=5.0,<6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Authentication & Security
django-allauth>=0.61.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2

# API & HTMX
django-htmx>=1.17.0

# Payments
stripe>=8.0.0

# File handling & Media
Pillow>=10.2.0
django-storages>=1.14.2
boto3>=1.34.0

# Background tasks
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.5.0

# Development
django-debug-toolbar>=4.3.0
django-extensions>=3.2.3

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
factory-boy>=3.3.0
coverage>=7.4.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Video meetings (optional)
# zoom-python-sdk or similar depending on choice

# Utilities
python-dateutil>=2.8.2
